+++ notification-service/src/main/java/n11bootcamp_project_backend/notification_service/config/RabbitMQConfig.java
@@ -1,0 +1,46 @@
+package n11bootcamp_project_backend.notification_service.config;
+
+import org.springframework.amqp.core.Binding;
+import org.springframework.amqp.core.BindingBuilder;
+import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.core.TopicExchange;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class RabbitMQConfig {
+
+    @Bean
+    public TopicExchange sagaExchange() {
+        return new TopicExchange("saga.exchange");
+    }
+
+    // payment.success eventini dinleyecek queue
+    @Bean
+    public Queue notificationQueue() {
+        return new Queue("notification.queue", true);
+    }
+
+    // Queue'yu exchange'e bağla
+    @Bean
+    public Binding notificationBinding(Queue notificationQueue,
+                                       TopicExchange sagaExchange) {
+        return BindingBuilder
+                .bind(notificationQueue)
+                .to(sagaExchange)
+                .with("payment.success");
+    }
+
+    @Bean
+    public Queue welcomeQueue() {
+        return new Queue("welcome.queue", true);
+    }
+
+    @Bean
+    public Binding welcomeBinding(Queue welcomeQueue, TopicExchange sagaExchange) {
+        return BindingBuilder
+                .bind(welcomeQueue)
+                .to(sagaExchange)
+                .with("user.registered");
+    }
+}