+++ notification-service/src/test/java/n11bootcamp_project_backend/notification_service/NotificationServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package n11bootcamp_project_backend.notification_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NotificationServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}